+++ brennfleck/jarod/tmr/scripts/events/ChatListener.java
@@ -1,8 +1,0 @@
-package brennfleck.jarod.tmr.scripts.events;
-
-import java.util.Observable;
-import java.util.Observer;
-
-public interface ChatListener extends TMRListener {
-	public void onChatEvent(ChatEvent e);
-}