+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-####-- Version 0.4 --
- + Added CHANGELOG to show all the changes! :D
- + Changed the Player class to be more unified with the Entity ways.
- + A bunch of JavaDocs added.
+####-- Version 0.5 --
+ - [=] <em>Player</em> is now called <em>ControlledPlayer</em>.
+  - [-] Removed the previous <em>Player</em> class.
+  - The player is mostly static, however getting locations you must use the getPlayer() method.
+ - [+] A way of accessing entities in the world.
+ - [+] Get the closest entity to a precise location.
+ - [+] You can now query what another player is doing.
+ - [+] Added an event listening system.
+  - [-] You can add event listeners, but nothing listens yet.
+
+######-- Version 0.4 --
+ - [+] Added CHANGELOG to show all the changes! :D
+ - [+] Changed the Player class to be more unified with the Entity ways.
+ - [+] A bunch of JavaDocs added.